Calculates distances between locations and sampling sites based on detection range.

Determine de euclidean distance from each occurrence to each survey (point), and mantain only the distances to each survey unit for occurrences inside a limiting distance of detection

pointSample(sites, occ, radius)

Arguments

sites

Data.frame with position (x, y) for each point survey (in rows).

occ

Data.frame with position (x, y) for each occurrence (in rows).

radius

A numeric value indicating the maximum distance of detection of the occurrence point in the survey.

Details

Value

Returns a list with a matrix with occ (rows) x sites (cols) with values of distance if less than radius, otherwise NA.

References

Author

Note

See also

Examples

obs <- data.frame(x=runif(1000, 0, 1), y=runif(1000, 0, 1))
sites <- data.frame(x=runif(10, 0, 1), y=runif(10, 0, 1))

psamp <- pointSample(obs, sites, 0.25)
psamp$distance
#>             [,1]       [,2]      [,3]      [,4]      [,5]       [,6]      [,7]
#>  [1,]         NA         NA 0.1725583 0.1486499 0.1109711 0.21031206        NA
#>            [,8]      [,9] [,10]      [,11]      [,12]     [,13]     [,14]
#>  [1,]        NA        NA    NA         NA         NA        NA        NA
#>           [,15]     [,16]      [,17]     [,18]     [,19] [,20]     [,21]
#>  [1,] 0.1286712        NA         NA        NA        NA    NA        NA
#>            [,22]      [,23]      [,24]    [,25]     [,26]      [,27]     [,28]
#>  [1,]         NA         NA         NA       NA        NA         NA        NA
#>            [,29]      [,30]      [,31]     [,32]     [,33]     [,34]     [,35]
#>  [1,]         NA         NA         NA        NA        NA        NA        NA
#>           [,36] [,37]      [,38]     [,39]     [,40] [,41]      [,42]     [,43]
#>  [1,]        NA    NA 0.21217758        NA        NA    NA         NA        NA
#>            [,44]      [,45]      [,46]     [,47] [,48]     [,49]     [,50]
#>  [1,]         NA         NA         NA        NA    NA        NA        NA
#>           [,51]      [,52]      [,53]     [,54]      [,55]      [,56]     [,57]
#>  [1,]        NA         NA         NA        NA         NA         NA        NA
#>           [,58] [,59]     [,60]     [,61]     [,62]      [,63] [,64]     [,65]
#>  [1,] 0.1359219    NA        NA 0.2167112        NA         NA    NA        NA
#>            [,66]     [,67]     [,68]     [,69]    [,70]      [,71]     [,72]
#>  [1,]         NA        NA        NA 0.0433907       NA         NA        NA
#>            [,73] [,74]     [,75]     [,76]      [,77]    [,78]      [,79]
#>  [1,] 0.17762863    NA        NA        NA         NA       NA         NA
#>           [,80]     [,81]     [,82]     [,83]      [,84]      [,85]     [,86]
#>  [1,]        NA        NA        NA 0.2034564         NA         NA        NA
#>          [,87]      [,88]     [,89]     [,90]    [,91]     [,92]      [,93]
#>  [1,]       NA         NA        NA        NA 0.125058        NA         NA
#>           [,94]     [,95] [,96]     [,97]     [,98]     [,99]   [,100]
#>  [1,]        NA        NA    NA        NA        NA        NA       NA
#>          [,101]    [,102]    [,103]    [,104]   [,105]    [,106]    [,107]
#>  [1,]        NA        NA 0.2225372        NA       NA        NA        NA
#>           [,108]    [,109]    [,110]    [,111]     [,112]    [,113] [,114]
#>  [1,]         NA 0.1368829 0.1396007        NA         NA 0.1677391     NA
#>          [,115]     [,116]     [,117]    [,118]     [,119]    [,120]    [,121]
#>  [1,]        NA 0.08933225         NA        NA         NA 0.1452163        NA
#>           [,122] [,123]    [,124]     [,125]    [,126] [,127]    [,128]
#>  [1,] 0.08471618     NA 0.1336372         NA        NA     NA        NA
#>           [,129]     [,130]    [,131]    [,132]    [,133]     [,134]     [,135]
#>  [1,]         NA         NA        NA        NA 0.2493866         NA         NA
#>           [,136]    [,137]    [,138]     [,139]    [,140]    [,141]    [,142]
#>  [1,]         NA        NA        NA         NA        NA        NA        NA
#>          [,143]     [,144]     [,145]   [,146]     [,147]     [,148]    [,149]
#>  [1,] 0.2401402 0.08079678 0.04664601       NA         NA 0.23656249        NA
#>          [,150]    [,151]     [,152]    [,153]    [,154]     [,155]    [,156]
#>  [1,]        NA        NA         NA        NA        NA         NA        NA
#>          [,157]    [,158]    [,159]    [,160]    [,161]     [,162]     [,163]
#>  [1,]        NA        NA        NA        NA        NA         NA         NA
#>          [,164]    [,165]    [,166]     [,167]    [,168]    [,169]    [,170]
#>  [1,]        NA        NA        NA         NA        NA 0.0931415        NA
#>          [,171]   [,172]    [,173]    [,174]     [,175]    [,176]    [,177]
#>  [1,] 0.2003286       NA 0.1453942        NA         NA        NA        NA
#>          [,178] [,179]    [,180]     [,181]    [,182]     [,183]    [,184]
#>  [1,]        NA     NA        NA         NA        NA 0.22848793        NA
#>          [,185]    [,186]    [,187] [,188]    [,189]    [,190]    [,191]
#>  [1,]        NA        NA        NA     NA        NA 0.1436959        NA
#>          [,192]    [,193]    [,194]    [,195]    [,196]    [,197]    [,198]
#>  [1,] 0.1781716        NA        NA        NA        NA        NA 0.2378439
#>           [,199]    [,200]    [,201] [,202]    [,203]    [,204]    [,205]
#>  [1,]         NA        NA        NA     NA        NA 0.1518810        NA
#>          [,206]   [,207]    [,208]    [,209] [,210]    [,211] [,212]     [,213]
#>  [1,]        NA       NA        NA        NA     NA        NA     NA         NA
#>          [,214]     [,215] [,216]    [,217]    [,218]    [,219]     [,220]
#>  [1,]        NA         NA     NA        NA 0.1170764        NA         NA
#>          [,221]    [,222] [,223]    [,224]    [,225] [,226]     [,227] [,228]
#>  [1,]        NA        NA     NA        NA        NA     NA         NA     NA
#>           [,229]    [,230]     [,231]     [,232]     [,233]    [,234]
#>  [1,]         NA        NA         NA         NA         NA 0.1836076
#>           [,235]    [,236] [,237]    [,238]    [,239]    [,240]    [,241]
#>  [1,]         NA        NA     NA        NA        NA        NA        NA
#>         [,242]    [,243]     [,244]    [,245] [,246]     [,247]    [,248]
#>  [1,]       NA        NA         NA        NA     NA         NA        NA
#>           [,249]    [,250]    [,251]    [,252]    [,253]    [,254]    [,255]
#>  [1,]         NA        NA 0.1364574        NA        NA        NA        NA
#>           [,256]    [,257]    [,258]     [,259] [,260]    [,261]    [,262]
#>  [1,]         NA 0.2425262        NA         NA     NA        NA        NA
#>          [,263]    [,264]    [,265]    [,266]    [,267] [,268]     [,269]
#>  [1,]        NA        NA        NA        NA        NA     NA 0.24771224
#>          [,270]     [,271]    [,272] [,273]    [,274] [,275]    [,276] [,277]
#>  [1,]        NA         NA        NA     NA        NA     NA        NA     NA
#>          [,278]    [,279]    [,280] [,281]    [,282]     [,283]    [,284]
#>  [1,]        NA        NA        NA     NA        NA         NA        NA
#>           [,285]    [,286]    [,287]     [,288]     [,289]    [,290]     [,291]
#>  [1,] 0.19340624        NA        NA         NA 0.04902039        NA         NA
#>          [,292]    [,293]    [,294]     [,295]    [,296]    [,297]    [,298]
#>  [1,]        NA 0.2247548        NA         NA        NA        NA        NA
#>           [,299]    [,300]    [,301]     [,302]    [,303]     [,304]     [,305]
#>  [1,]         NA        NA        NA         NA        NA 0.06273524         NA
#>          [,306]    [,307]    [,308]    [,309]     [,310]    [,311]    [,312]
#>  [1,] 0.2433365        NA        NA        NA         NA        NA        NA
#>          [,313] [,314]    [,315]    [,316]    [,317]     [,318]    [,319]
#>  [1,]        NA     NA        NA        NA        NA         NA        NA
#>       [,320]    [,321]    [,322]     [,323]    [,324]    [,325]    [,326]
#>  [1,]     NA        NA        NA         NA        NA        NA        NA
#>           [,327]    [,328] [,329]    [,330]    [,331]    [,332]     [,333]
#>  [1,] 0.08698391 0.1301274     NA 0.0154464        NA        NA         NA
#>          [,334]    [,335]    [,336]    [,337]     [,338]    [,339]    [,340]
#>  [1,]        NA 0.1676828        NA        NA         NA        NA        NA
#>          [,341]     [,342]    [,343]     [,344]     [,345]     [,346] [,347]
#>  [1,]        NA         NA        NA         NA         NA         NA     NA
#>       [,348]    [,349]    [,350]     [,351]     [,352]    [,353]     [,354]
#>  [1,]     NA        NA        NA         NA         NA        NA 0.16629303
#>          [,355]    [,356]     [,357]    [,358]    [,359]     [,360]    [,361]
#>  [1,]        NA        NA 0.09104083        NA        NA         NA        NA
#>           [,362]    [,363]    [,364]   [,365]    [,366]   [,367]    [,368]
#>  [1,]         NA        NA        NA       NA        NA       NA        NA
#>       [,369]    [,370] [,371]    [,372]    [,373]    [,374]     [,375]
#>  [1,]     NA        NA     NA        NA        NA        NA         NA
#>          [,376] [,377]    [,378]    [,379]    [,380]    [,381] [,382]    [,383]
#>  [1,]        NA     NA        NA        NA        NA        NA     NA        NA
#>           [,384]    [,385]    [,386]    [,387]    [,388]    [,389]     [,390]
#>  [1,]         NA        NA 0.2160531        NA        NA        NA         NA
#>          [,391]   [,392]    [,393] [,394]    [,395]    [,396]    [,397]
#>  [1,]        NA       NA        NA     NA        NA 0.1794921        NA
#>          [,398]   [,399]   [,400]    [,401]    [,402]    [,403]    [,404]
#>  [1,]        NA       NA       NA        NA        NA        NA        NA
#>          [,405]    [,406]    [,407]    [,408]    [,409]    [,410]    [,411]
#>  [1,] 0.2397293        NA        NA        NA        NA        NA        NA
#>           [,412]    [,413]     [,414]    [,415]    [,416]    [,417] [,418]
#>  [1,]         NA        NA         NA        NA        NA        NA     NA
#>          [,419]     [,420]    [,421]    [,422]    [,423]    [,424]    [,425]
#>  [1,]        NA         NA        NA        NA 0.1710799        NA        NA
#>         [,426]     [,427]    [,428]     [,429]    [,430] [,431]    [,432]
#>  [1,] 0.101723         NA        NA         NA        NA     NA        NA
#>           [,433] [,434]     [,435] [,436]    [,437] [,438]     [,439]
#>  [1,] 0.05740498     NA         NA     NA        NA     NA         NA
#>           [,440]    [,441]    [,442]    [,443]    [,444]     [,445]    [,446]
#>  [1,]         NA        NA 0.1729778        NA 0.1952807         NA        NA
#>          [,447]     [,448]    [,449]     [,450]     [,451]    [,452]     [,453]
#>  [1,]        NA         NA        NA         NA 0.02507756 0.1314954         NA
#>           [,454]     [,455] [,456]     [,457]   [,458]    [,459]    [,460]
#>  [1,]         NA         NA     NA         NA       NA        NA        NA
#>          [,461]     [,462]    [,463]    [,464]     [,465]    [,466]    [,467]
#>  [1,] 0.2133792         NA        NA        NA         NA        NA        NA
#>          [,468]    [,469]     [,470]     [,471]    [,472]    [,473]    [,474]
#>  [1,]        NA        NA 0.08690294         NA 0.2285616 0.1232647        NA
#>          [,475]     [,476]    [,477]   [,478]     [,479]     [,480]    [,481]
#>  [1,]        NA         NA        NA       NA         NA         NA        NA
#>       [,482]    [,483]    [,484]     [,485]    [,486]     [,487]     [,488]
#>  [1,]     NA        NA        NA 0.06878971 0.1964560         NA         NA
#>          [,489]     [,490]    [,491]    [,492]    [,493]    [,494]    [,495]
#>  [1,]        NA 0.24147788        NA        NA        NA        NA        NA
#>          [,496]     [,497]    [,498] [,499] [,500]     [,501]    [,502]
#>  [1,] 0.1196584         NA        NA     NA     NA         NA 0.2001766
#>          [,503]    [,504]    [,505]    [,506] [,507]    [,508]     [,509]
#>  [1,] 0.2053250 0.1366182 0.2179272        NA     NA        NA         NA
#>           [,510]    [,511]     [,512]     [,513]    [,514]      [,515]
#>  [1,]         NA        NA         NA         NA 0.2131200          NA
#>           [,516]    [,517]    [,518]    [,519]     [,520]     [,521]    [,522]
#>  [1,]         NA        NA        NA        NA         NA         NA        NA
#>         [,523]     [,524]    [,525]    [,526]     [,527]     [,528]    [,529]
#>  [1,]       NA         NA        NA 0.1579403         NA 0.08237719        NA
#>          [,530]    [,531]     [,532]    [,533] [,534]    [,535]    [,536]
#>  [1,]        NA 0.1162797         NA        NA     NA        NA        NA
#>           [,537]    [,538]    [,539]    [,540]    [,541]     [,542]    [,543]
#>  [1,]         NA        NA 0.2299003        NA        NA         NA 0.2061532
#>          [,544]    [,545]    [,546]     [,547]     [,548]    [,549]     [,550]
#>  [1,] 0.1088865        NA        NA         NA         NA        NA         NA
#>           [,551]     [,552]     [,553]    [,554]    [,555]    [,556]     [,557]
#>  [1,] 0.05303807         NA 0.06401315        NA        NA 0.1183429         NA
#>          [,558]    [,559]    [,560]    [,561]     [,562]    [,563]     [,564]
#>  [1,]        NA        NA        NA        NA         NA        NA         NA
#>          [,565] [,566]    [,567]    [,568] [,569]    [,570]    [,571]    [,572]
#>  [1,]        NA     NA        NA        NA     NA 0.1913300        NA        NA
#>          [,573]   [,574]     [,575]    [,576]     [,577]     [,578]     [,579]
#>  [1,]        NA       NA         NA        NA         NA         NA 0.22593342
#>          [,580]     [,581]    [,582]   [,583]    [,584]    [,585]     [,586]
#>  [1,]        NA         NA        NA       NA        NA        NA         NA
#>       [,587]    [,588]    [,589]    [,590]    [,591]     [,592]    [,593]
#>  [1,]     NA        NA        NA        NA        NA         NA        NA
#>          [,594]     [,595]    [,596]    [,597]      [,598]    [,599]    [,600]
#>  [1,]        NA         NA 0.2419734 0.1213369 0.003820795        NA        NA
#>          [,601]    [,602]    [,603]    [,604]    [,605]    [,606] [,607]
#>  [1,]        NA        NA        NA        NA        NA        NA     NA
#>         [,608]    [,609]    [,610]    [,611]    [,612]    [,613]     [,614]
#>  [1,]       NA        NA        NA        NA        NA        NA         NA
#>          [,615]    [,616]     [,617]     [,618]    [,619]    [,620] [,621]
#>  [1,]        NA        NA         NA 0.09504786 0.1751223 0.2166334     NA
#>          [,622]     [,623]    [,624]    [,625]    [,626] [,627]     [,628]
#>  [1,]        NA 0.07849515 0.1491911        NA        NA     NA         NA
#>       [,629]    [,630]     [,631]    [,632]    [,633]     [,634]    [,635]
#>  [1,]     NA 0.1297491         NA        NA        NA         NA        NA
#>       [,636]     [,637]    [,638]    [,639]     [,640]    [,641] [,642]
#>  [1,]     NA 0.06622822        NA        NA         NA        NA     NA
#>          [,643]    [,644]    [,645]     [,646] [,647]    [,648]    [,649]
#>  [1,] 0.0449348        NA        NA         NA     NA        NA        NA
#>          [,650]    [,651] [,652] [,653]    [,654] [,655]    [,656]    [,657]
#>  [1,]        NA        NA     NA     NA        NA     NA        NA        NA
#>         [,658]    [,659] [,660]    [,661]    [,662]     [,663] [,664]
#>  [1,]       NA        NA     NA        NA        NA 0.18086896     NA
#>           [,665]    [,666]    [,667]     [,668]     [,669]  [,670]    [,671]
#>  [1,]         NA 0.1691320 0.1696378         NA         NA      NA 0.2371022
#>          [,672]     [,673]     [,674]    [,675] [,676]     [,677]   [,678]
#>  [1,]        NA         NA         NA        NA     NA         NA       NA
#>       [,679]    [,680]     [,681]    [,682] [,683]    [,684]    [,685]
#>  [1,]     NA 0.0833134         NA 0.1805540     NA        NA        NA
#>          [,686]    [,687]     [,688]    [,689]    [,690]    [,691]    [,692]
#>  [1,] 0.1211599        NA 0.09410965 0.2117079        NA        NA        NA
#>          [,693] [,694]     [,695] [,696]    [,697]    [,698]    [,699]
#>  [1,]        NA     NA         NA     NA        NA        NA        NA
#>           [,700]    [,701]    [,702]     [,703]     [,704]      [,705]
#>  [1,] 0.04967847        NA        NA         NA         NA          NA
#>          [,706] [,707]    [,708]    [,709]    [,710] [,711]    [,712]    [,713]
#>  [1,]        NA     NA        NA        NA        NA     NA 0.1154965 0.1942095
#>        [,714]    [,715]    [,716]    [,717]    [,718]    [,719] [,720] [,721]
#>  [1,]      NA        NA        NA        NA        NA        NA     NA     NA
#>       [,722]    [,723] [,724]    [,725]    [,726] [,727]    [,728]    [,729]
#>  [1,]     NA        NA     NA        NA        NA     NA 0.1544276        NA
#>           [,730]    [,731]    [,732]     [,733]    [,734]    [,735]    [,736]
#>  [1,]         NA        NA 0.2199669         NA        NA        NA        NA
#>          [,737]    [,738]    [,739]    [,740]    [,741]    [,742]    [,743]
#>  [1,]        NA 0.1464235        NA        NA        NA        NA        NA
#>          [,744]   [,745]     [,746]    [,747] [,748]    [,749]     [,750]
#>  [1,]        NA       NA         NA        NA     NA        NA         NA
#>          [,751] [,752]    [,753]    [,754]    [,755]     [,756]   [,757]
#>  [1,]        NA     NA        NA        NA        NA 0.08767582 0.161095
#>          [,758]    [,759] [,760] [,761]    [,762]    [,763]     [,764]
#>  [1,]        NA        NA     NA     NA        NA 0.2136176         NA
#>          [,765]    [,766]     [,767]    [,768]     [,769]    [,770]     [,771]
#>  [1,]        NA 0.1348524         NA        NA         NA        NA         NA
#>           [,772]     [,773]    [,774]     [,775]    [,776]    [,777]     [,778]
#>  [1,]         NA         NA        NA 0.24945752 0.2365417        NA 0.07099946
#>       [,779]    [,780]    [,781]     [,782]   [,783]    [,784]    [,785]
#>  [1,]     NA        NA        NA         NA       NA        NA 0.1530521
#>           [,786]    [,787]     [,788]    [,789]    [,790]     [,791]    [,792]
#>  [1,]         NA        NA         NA        NA 0.1417155         NA        NA
#>          [,793]    [,794]     [,795]    [,796]    [,797]     [,798]    [,799]
#>  [1,]        NA        NA         NA        NA        NA         NA        NA
#>          [,800]     [,801]    [,802]     [,803]     [,804]    [,805]    [,806]
#>  [1,] 0.1478355         NA        NA         NA 0.24886199        NA        NA
#>          [,807]    [,808]     [,809]    [,810]    [,811]     [,812]    [,813]
#>  [1,]        NA        NA 0.08012195        NA        NA         NA        NA
#>           [,814]    [,815]    [,816]    [,817]     [,818]    [,819]     [,820]
#>  [1,] 0.02320153        NA        NA        NA         NA        NA         NA
#>         [,821]     [,822]    [,823]    [,824]     [,825]    [,826] [,827]
#>  [1,]       NA         NA        NA        NA 0.18620374 0.1794064     NA
#>          [,828]     [,829] [,830]    [,831]    [,832]    [,833]    [,834]
#>  [1,]        NA         NA     NA 0.1321229        NA 0.1350174 0.1735907
#>          [,835]    [,836]     [,837]   [,838]     [,839]    [,840]     [,841]
#>  [1,]        NA        NA         NA       NA         NA        NA 0.05187109
#>          [,842]     [,843]    [,844]    [,845]    [,846]     [,847]    [,848]
#>  [1,]        NA         NA        NA        NA        NA         NA        NA
#>          [,849]     [,850] [,851]    [,852]    [,853]    [,854]    [,855]
#>  [1,] 0.1609019         NA     NA        NA        NA        NA        NA
#>           [,856]    [,857]    [,858]    [,859]    [,860]    [,861]     [,862]
#>  [1,]         NA        NA        NA        NA        NA        NA         NA
#>           [,863]    [,864]    [,865]    [,866]    [,867]    [,868]    [,869]
#>  [1,]         NA        NA        NA        NA        NA        NA 0.1217702
#>           [,870]    [,871]     [,872]    [,873]    [,874]     [,875]    [,876]
#>  [1,]         NA        NA         NA        NA 0.2058873         NA        NA
#>          [,877]     [,878]    [,879]     [,880]     [,881]    [,882]    [,883]
#>  [1,]        NA         NA        NA         NA         NA        NA        NA
#>          [,884]     [,885]    [,886]     [,887]   [,888]    [,889]    [,890]
#>  [1,] 0.1795518 0.20032792 0.1792041         NA       NA        NA        NA
#>           [,891]    [,892]    [,893]     [,894] [,895]    [,896]    [,897]
#>  [1,]         NA        NA        NA 0.06446138     NA        NA        NA
#>         [,898]    [,899] [,900] [,901]     [,902]     [,903]    [,904]
#>  [1,]       NA        NA     NA     NA         NA 0.23125272        NA
#>          [,905]    [,906]     [,907]     [,908]     [,909] [,910]    [,911]
#>  [1,]        NA        NA         NA         NA         NA     NA        NA
#>       [,912] [,913]     [,914]    [,915]    [,916]    [,917] [,918]    [,919]
#>  [1,]     NA     NA 0.08492683        NA        NA        NA     NA        NA
#>           [,920]     [,921]    [,922]    [,923]     [,924]    [,925]    [,926]
#>  [1,]         NA 0.09567478        NA 0.1236042         NA        NA 0.1193506
#>           [,927]    [,928]    [,929]    [,930]     [,931]    [,932]    [,933]
#>  [1,]         NA 0.2280755 0.2354543        NA 0.05223047 0.1579349        NA
#>          [,934]    [,935] [,936]     [,937]    [,938]    [,939]    [,940]
#>  [1,]        NA 0.1028610     NA 0.24983189        NA        NA        NA
#>           [,941]    [,942]    [,943]    [,944]    [,945]    [,946]   [,947]
#>  [1,] 0.09042887        NA        NA        NA        NA        NA       NA
#>          [,948]    [,949]    [,950]    [,951]     [,952]    [,953]     [,954]
#>  [1,]        NA        NA        NA        NA         NA 0.1601628 0.09509606
#>       [,955]    [,956]    [,957] [,958]    [,959]   [,960]     [,961]
#>  [1,]     NA        NA        NA     NA 0.1386323       NA         NA
#>           [,962]     [,963]    [,964] [,965]     [,966]    [,967]    [,968]
#>  [1,]         NA 0.02236027        NA     NA         NA        NA        NA
#>          [,969] [,970]     [,971]   [,972]    [,973] [,974]    [,975]
#>  [1,]        NA     NA         NA       NA        NA     NA 0.1602615
#>           [,976]    [,977]    [,978] [,979] [,980] [,981] [,982]    [,983]
#>  [1,]         NA        NA        NA     NA     NA     NA     NA        NA
#>          [,984]     [,985]    [,986]     [,987]    [,988] [,989]     [,990]
#>  [1,]        NA         NA 0.1162956         NA        NA     NA         NA
#>          [,991]    [,992]    [,993]    [,994] [,995]     [,996]    [,997]
#>  [1,]        NA 0.2469056        NA        NA     NA         NA        NA
#>          [,998]    [,999]    [,1000]
#>  [1,]        NA        NA         NA
#>  [ reached 'max' / getOption("max.print") -- omitted 9 rows ]