Create sampling points based on uniform sampling grid centroids.

Establish sampling points as a grid defined by X and Y ranges and a specified cell size.

uniformSamp(xrange, yrange, cellsize)

Arguments

xrange

Vector of two elements idnicating x coordinates range (min,max).

yrange

Same as above for y range.

cellsize

Cell size for grid (minimum dist between points)

Details

Value

Returns a table of coordinates of the sampling sites.

References

Author

Note

See also

Examples

# Create 25 sampling sites in a square grid.
sampling <- uniformSamp(c(0,100), c(0,100), 5)
sampling
#>        x    y
#> 1    2.5  2.5
#> 2    5.0  2.5
#> 3    7.5  2.5
#> 4   10.0  2.5
#> 5   12.5  2.5
#> 6   15.0  2.5
#> 7   17.5  2.5
#> 8   20.0  2.5
#> 9   22.5  2.5
#> 10  25.0  2.5
#> 11  27.5  2.5
#> 12  30.0  2.5
#> 13  32.5  2.5
#> 14  35.0  2.5
#> 15  37.5  2.5
#> 16  40.0  2.5
#> 17  42.5  2.5
#> 18  45.0  2.5
#> 19  47.5  2.5
#> 20  50.0  2.5
#> 21  52.5  2.5
#> 22  55.0  2.5
#> 23  57.5  2.5
#> 24  60.0  2.5
#> 25  62.5  2.5
#> 26  65.0  2.5
#> 27  67.5  2.5
#> 28  70.0  2.5
#> 29  72.5  2.5
#> 30  75.0  2.5
#> 31  77.5  2.5
#> 32  80.0  2.5
#> 33  82.5  2.5
#> 34  85.0  2.5
#> 35  87.5  2.5
#> 36  90.0  2.5
#> 37  92.5  2.5
#> 38  95.0  2.5
#> 39  97.5  2.5
#> 40   2.5  5.0
#> 41   5.0  5.0
#> 42   7.5  5.0
#> 43  10.0  5.0
#> 44  12.5  5.0
#> 45  15.0  5.0
#> 46  17.5  5.0
#> 47  20.0  5.0
#> 48  22.5  5.0
#> 49  25.0  5.0
#> 50  27.5  5.0
#> 51  30.0  5.0
#> 52  32.5  5.0
#> 53  35.0  5.0
#> 54  37.5  5.0
#> 55  40.0  5.0
#> 56  42.5  5.0
#> 57  45.0  5.0
#> 58  47.5  5.0
#> 59  50.0  5.0
#> 60  52.5  5.0
#> 61  55.0  5.0
#> 62  57.5  5.0
#> 63  60.0  5.0
#> 64  62.5  5.0
#> 65  65.0  5.0
#> 66  67.5  5.0
#> 67  70.0  5.0
#> 68  72.5  5.0
#> 69  75.0  5.0
#> 70  77.5  5.0
#> 71  80.0  5.0
#> 72  82.5  5.0
#> 73  85.0  5.0
#> 74  87.5  5.0
#> 75  90.0  5.0
#> 76  92.5  5.0
#> 77  95.0  5.0
#> 78  97.5  5.0
#> 79   2.5  7.5
#> 80   5.0  7.5
#> 81   7.5  7.5
#> 82  10.0  7.5
#> 83  12.5  7.5
#> 84  15.0  7.5
#> 85  17.5  7.5
#> 86  20.0  7.5
#> 87  22.5  7.5
#> 88  25.0  7.5
#> 89  27.5  7.5
#> 90  30.0  7.5
#> 91  32.5  7.5
#> 92  35.0  7.5
#> 93  37.5  7.5
#> 94  40.0  7.5
#> 95  42.5  7.5
#> 96  45.0  7.5
#> 97  47.5  7.5
#> 98  50.0  7.5
#> 99  52.5  7.5
#> 100 55.0  7.5
#> 101 57.5  7.5
#> 102 60.0  7.5
#> 103 62.5  7.5
#> 104 65.0  7.5
#> 105 67.5  7.5
#> 106 70.0  7.5
#> 107 72.5  7.5
#> 108 75.0  7.5
#> 109 77.5  7.5
#> 110 80.0  7.5
#> 111 82.5  7.5
#> 112 85.0  7.5
#> 113 87.5  7.5
#> 114 90.0  7.5
#> 115 92.5  7.5
#> 116 95.0  7.5
#> 117 97.5  7.5
#> 118  2.5 10.0
#> 119  5.0 10.0
#> 120  7.5 10.0
#> 121 10.0 10.0
#> 122 12.5 10.0
#> 123 15.0 10.0
#> 124 17.5 10.0
#> 125 20.0 10.0
#> 126 22.5 10.0
#> 127 25.0 10.0
#> 128 27.5 10.0
#> 129 30.0 10.0
#> 130 32.5 10.0
#> 131 35.0 10.0
#> 132 37.5 10.0
#> 133 40.0 10.0
#> 134 42.5 10.0
#> 135 45.0 10.0
#> 136 47.5 10.0
#> 137 50.0 10.0
#> 138 52.5 10.0
#> 139 55.0 10.0
#> 140 57.5 10.0
#> 141 60.0 10.0
#> 142 62.5 10.0
#> 143 65.0 10.0
#> 144 67.5 10.0
#> 145 70.0 10.0
#> 146 72.5 10.0
#> 147 75.0 10.0
#> 148 77.5 10.0
#> 149 80.0 10.0
#> 150 82.5 10.0
#> 151 85.0 10.0
#> 152 87.5 10.0
#> 153 90.0 10.0
#> 154 92.5 10.0
#> 155 95.0 10.0
#> 156 97.5 10.0
#> 157  2.5 12.5
#> 158  5.0 12.5
#> 159  7.5 12.5
#> 160 10.0 12.5
#> 161 12.5 12.5
#> 162 15.0 12.5
#> 163 17.5 12.5
#> 164 20.0 12.5
#> 165 22.5 12.5
#> 166 25.0 12.5
#> 167 27.5 12.5
#> 168 30.0 12.5
#> 169 32.5 12.5
#> 170 35.0 12.5
#> 171 37.5 12.5
#> 172 40.0 12.5
#> 173 42.5 12.5
#> 174 45.0 12.5
#> 175 47.5 12.5
#> 176 50.0 12.5
#> 177 52.5 12.5
#> 178 55.0 12.5
#> 179 57.5 12.5
#> 180 60.0 12.5
#> 181 62.5 12.5
#> 182 65.0 12.5
#> 183 67.5 12.5
#> 184 70.0 12.5
#> 185 72.5 12.5
#> 186 75.0 12.5
#> 187 77.5 12.5
#> 188 80.0 12.5
#> 189 82.5 12.5
#> 190 85.0 12.5
#> 191 87.5 12.5
#> 192 90.0 12.5
#> 193 92.5 12.5
#> 194 95.0 12.5
#> 195 97.5 12.5
#> 196  2.5 15.0
#> 197  5.0 15.0
#> 198  7.5 15.0
#> 199 10.0 15.0
#> 200 12.5 15.0
#> 201 15.0 15.0
#> 202 17.5 15.0
#> 203 20.0 15.0
#> 204 22.5 15.0
#> 205 25.0 15.0
#> 206 27.5 15.0
#> 207 30.0 15.0
#> 208 32.5 15.0
#> 209 35.0 15.0
#> 210 37.5 15.0
#> 211 40.0 15.0
#> 212 42.5 15.0
#> 213 45.0 15.0
#> 214 47.5 15.0
#> 215 50.0 15.0
#> 216 52.5 15.0
#> 217 55.0 15.0
#> 218 57.5 15.0
#> 219 60.0 15.0
#> 220 62.5 15.0
#> 221 65.0 15.0
#> 222 67.5 15.0
#> 223 70.0 15.0
#> 224 72.5 15.0
#> 225 75.0 15.0
#> 226 77.5 15.0
#> 227 80.0 15.0
#> 228 82.5 15.0
#> 229 85.0 15.0
#> 230 87.5 15.0
#> 231 90.0 15.0
#> 232 92.5 15.0
#> 233 95.0 15.0
#> 234 97.5 15.0
#> 235  2.5 17.5
#> 236  5.0 17.5
#> 237  7.5 17.5
#> 238 10.0 17.5
#> 239 12.5 17.5
#> 240 15.0 17.5
#> 241 17.5 17.5
#> 242 20.0 17.5
#> 243 22.5 17.5
#> 244 25.0 17.5
#> 245 27.5 17.5
#> 246 30.0 17.5
#> 247 32.5 17.5
#> 248 35.0 17.5
#> 249 37.5 17.5
#> 250 40.0 17.5
#> 251 42.5 17.5
#> 252 45.0 17.5
#> 253 47.5 17.5
#> 254 50.0 17.5
#> 255 52.5 17.5
#> 256 55.0 17.5
#> 257 57.5 17.5
#> 258 60.0 17.5
#> 259 62.5 17.5
#> 260 65.0 17.5
#> 261 67.5 17.5
#> 262 70.0 17.5
#> 263 72.5 17.5
#> 264 75.0 17.5
#> 265 77.5 17.5
#> 266 80.0 17.5
#> 267 82.5 17.5
#> 268 85.0 17.5
#> 269 87.5 17.5
#> 270 90.0 17.5
#> 271 92.5 17.5
#> 272 95.0 17.5
#> 273 97.5 17.5
#> 274  2.5 20.0
#> 275  5.0 20.0
#> 276  7.5 20.0
#> 277 10.0 20.0
#> 278 12.5 20.0
#> 279 15.0 20.0
#> 280 17.5 20.0
#> 281 20.0 20.0
#> 282 22.5 20.0
#> 283 25.0 20.0
#> 284 27.5 20.0
#> 285 30.0 20.0
#> 286 32.5 20.0
#> 287 35.0 20.0
#> 288 37.5 20.0
#> 289 40.0 20.0
#> 290 42.5 20.0
#> 291 45.0 20.0
#> 292 47.5 20.0
#> 293 50.0 20.0
#> 294 52.5 20.0
#> 295 55.0 20.0
#> 296 57.5 20.0
#> 297 60.0 20.0
#> 298 62.5 20.0
#> 299 65.0 20.0
#> 300 67.5 20.0
#> 301 70.0 20.0
#> 302 72.5 20.0
#> 303 75.0 20.0
#> 304 77.5 20.0
#> 305 80.0 20.0
#> 306 82.5 20.0
#> 307 85.0 20.0
#> 308 87.5 20.0
#> 309 90.0 20.0
#> 310 92.5 20.0
#> 311 95.0 20.0
#> 312 97.5 20.0
#> 313  2.5 22.5
#> 314  5.0 22.5
#> 315  7.5 22.5
#> 316 10.0 22.5
#> 317 12.5 22.5
#> 318 15.0 22.5
#> 319 17.5 22.5
#> 320 20.0 22.5
#> 321 22.5 22.5
#> 322 25.0 22.5
#> 323 27.5 22.5
#> 324 30.0 22.5
#> 325 32.5 22.5
#> 326 35.0 22.5
#> 327 37.5 22.5
#> 328 40.0 22.5
#> 329 42.5 22.5
#> 330 45.0 22.5
#> 331 47.5 22.5
#> 332 50.0 22.5
#> 333 52.5 22.5
#> 334 55.0 22.5
#> 335 57.5 22.5
#> 336 60.0 22.5
#> 337 62.5 22.5
#> 338 65.0 22.5
#> 339 67.5 22.5
#> 340 70.0 22.5
#> 341 72.5 22.5
#> 342 75.0 22.5
#> 343 77.5 22.5
#> 344 80.0 22.5
#> 345 82.5 22.5
#> 346 85.0 22.5
#> 347 87.5 22.5
#> 348 90.0 22.5
#> 349 92.5 22.5
#> 350 95.0 22.5
#> 351 97.5 22.5
#> 352  2.5 25.0
#> 353  5.0 25.0
#> 354  7.5 25.0
#> 355 10.0 25.0
#> 356 12.5 25.0
#> 357 15.0 25.0
#> 358 17.5 25.0
#> 359 20.0 25.0
#> 360 22.5 25.0
#> 361 25.0 25.0
#> 362 27.5 25.0
#> 363 30.0 25.0
#> 364 32.5 25.0
#> 365 35.0 25.0
#> 366 37.5 25.0
#> 367 40.0 25.0
#> 368 42.5 25.0
#> 369 45.0 25.0
#> 370 47.5 25.0
#> 371 50.0 25.0
#> 372 52.5 25.0
#> 373 55.0 25.0
#> 374 57.5 25.0
#> 375 60.0 25.0
#> 376 62.5 25.0
#> 377 65.0 25.0
#> 378 67.5 25.0
#> 379 70.0 25.0
#> 380 72.5 25.0
#> 381 75.0 25.0
#> 382 77.5 25.0
#> 383 80.0 25.0
#> 384 82.5 25.0
#> 385 85.0 25.0
#> 386 87.5 25.0
#> 387 90.0 25.0
#> 388 92.5 25.0
#> 389 95.0 25.0
#> 390 97.5 25.0
#> 391  2.5 27.5
#> 392  5.0 27.5
#> 393  7.5 27.5
#> 394 10.0 27.5
#> 395 12.5 27.5
#> 396 15.0 27.5
#> 397 17.5 27.5
#> 398 20.0 27.5
#> 399 22.5 27.5
#> 400 25.0 27.5
#> 401 27.5 27.5
#> 402 30.0 27.5
#> 403 32.5 27.5
#> 404 35.0 27.5
#> 405 37.5 27.5
#> 406 40.0 27.5
#> 407 42.5 27.5
#> 408 45.0 27.5
#> 409 47.5 27.5
#> 410 50.0 27.5
#> 411 52.5 27.5
#> 412 55.0 27.5
#> 413 57.5 27.5
#> 414 60.0 27.5
#> 415 62.5 27.5
#> 416 65.0 27.5
#> 417 67.5 27.5
#> 418 70.0 27.5
#> 419 72.5 27.5
#> 420 75.0 27.5
#> 421 77.5 27.5
#> 422 80.0 27.5
#> 423 82.5 27.5
#> 424 85.0 27.5
#> 425 87.5 27.5
#> 426 90.0 27.5
#> 427 92.5 27.5
#> 428 95.0 27.5
#> 429 97.5 27.5
#> 430  2.5 30.0
#> 431  5.0 30.0
#> 432  7.5 30.0
#> 433 10.0 30.0
#> 434 12.5 30.0
#> 435 15.0 30.0
#> 436 17.5 30.0
#> 437 20.0 30.0
#> 438 22.5 30.0
#> 439 25.0 30.0
#> 440 27.5 30.0
#> 441 30.0 30.0
#> 442 32.5 30.0
#> 443 35.0 30.0
#> 444 37.5 30.0
#> 445 40.0 30.0
#> 446 42.5 30.0
#> 447 45.0 30.0
#> 448 47.5 30.0
#> 449 50.0 30.0
#> 450 52.5 30.0
#> 451 55.0 30.0
#> 452 57.5 30.0
#> 453 60.0 30.0
#> 454 62.5 30.0
#> 455 65.0 30.0
#> 456 67.5 30.0
#> 457 70.0 30.0
#> 458 72.5 30.0
#> 459 75.0 30.0
#> 460 77.5 30.0
#> 461 80.0 30.0
#> 462 82.5 30.0
#> 463 85.0 30.0
#> 464 87.5 30.0
#> 465 90.0 30.0
#> 466 92.5 30.0
#> 467 95.0 30.0
#> 468 97.5 30.0
#> 469  2.5 32.5
#> 470  5.0 32.5
#> 471  7.5 32.5
#> 472 10.0 32.5
#> 473 12.5 32.5
#> 474 15.0 32.5
#> 475 17.5 32.5
#> 476 20.0 32.5
#> 477 22.5 32.5
#> 478 25.0 32.5
#> 479 27.5 32.5
#> 480 30.0 32.5
#> 481 32.5 32.5
#> 482 35.0 32.5
#> 483 37.5 32.5
#> 484 40.0 32.5
#> 485 42.5 32.5
#> 486 45.0 32.5
#> 487 47.5 32.5
#> 488 50.0 32.5
#> 489 52.5 32.5
#> 490 55.0 32.5
#> 491 57.5 32.5
#> 492 60.0 32.5
#> 493 62.5 32.5
#> 494 65.0 32.5
#> 495 67.5 32.5
#> 496 70.0 32.5
#> 497 72.5 32.5
#> 498 75.0 32.5
#> 499 77.5 32.5
#> 500 80.0 32.5
#>  [ reached 'max' / getOption("max.print") -- omitted 1021 rows ]